requirements-core.txt
# RUBE Core Dependencies — Required for basic operation
anthropic>=0.84.0
python-dotenv>=1.0.0
requests>=2.31.0
pygame>=2.5.0
psutil>=5.9.0
Pillow>=10.0.0
groq>=0.5.0
cartesia>=1.0.0
deepgram-sdk>=6.0.0
edge-tts>=6.1.0
sounddevice>=0.4.6
colorama>=0.4.6
